dev-libs/openssl: add USE `quic` to openssl
OpenSSL 3.2+ brings an initial RFC 9000 (QUIC)
implementation that is sufficient for use as a client
or for general experimentation and QUIC / HTTP/3
development.

While 3.2 is sufficient, the API exposed by 3.3 is
more mature; we'll only enable it from here so that
client applications (e.g. net-misc/curl) can be wired up
to test QUIC.

This passes tests upstream and is reported to work, but
Gentoo QUIC support is best described as 'developing';
hic sunt dracones.

Signed-off-by: Matt Jolly <kangie@gentoo.org>