dev-qt/qtwebengine: fix build with USE="pdfium -pulseaudio" (qt6)
Or hopefully anyway, have not tested the full build without libpulse,
but it at least no longer looks for it.

Unclear whether pdfium was automagically linking with it, or just
looking for it for nothing while unused. The former technically
needs a revbump, but not worth it given the long build times and
how pdfium is scarcely used.

Closes: https://bugs.gentoo.org/934635
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>