dev-qt/qtwebengine: alternate fix for USE="pdfium -pulseaudio" (qt6)
Just realized the issue is self-inflicted. The pkg-config check
is done only if "use_pulseaudio && link_pulseaudio" and the latter
is passed by us rather than Qt.

It seemed harmless to be unconditional given the main switch
disabled it (which technically sounds better), but given it
doesn't for pdfium let's change that and do it for pipewire
as well while at it.

Bug: https://bugs.gentoo.org/934635
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>