app-shells/bash: do not update internal window title of screen and tmux
The k escape sequence changes the internal window title of screen and
tmux, which is an alias given to the window by the user. The alias can
be used to reference the window in internal commands, e.g. by using
"select <alias>" in screen or "select-window -t <alias>" in tmux to
switch to a window by name, and should not be updated by commands
running inside of screen and tmux.

Closes: https://bugs.gentoo.org/934732
Signed-off-by: Sven Wegener <swegener@gentoo.org>