dev-python/flit: remove useless 4 year old dependencies
requests_download hasn't been used since a deprecated feature available
as a pip "extra" was removed:
https://github.com/pypa/flit/commit/b9a8ecd6037b3b98452c6fa90697612075c03ef4

Arguably before then it should have been an optfeature but at this point
it is pointless to discuss -- the feature got dropped in flit 3.0.0.

tomli has never been a dependency at all. The first version of flit that
moved from "toml" (an abandoned project) to tomli, only used tomli_w in
flit, and relied on flit_core to handle all use of "tomli" (and
flit-core correctly depends on the stdlib tomllib where reasonable).

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>