dev-util/pkgdev: add dependency on tomli for old impls
It is missing it in wheel metadata, and also in the ebuild metadata. It
is only necessary for versions of python that don't have it in the
stdlib, and all those are old versions, but that's no reason not to
depend correctly on it where needed.

The "doc" USE flag also depends on tomli and does so unconditionally. It
appears that it doesn't allow one to select specific USEDEPs, which is
interesting but ok.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>