dev-python/doit: conditionalize tomli/tomlkit imports after stdlib tomllib
This project attempts to support every kitchen sink when it comes to
loading a toml file. And since 0.36.0 it also supports the
dev-python/tomli library:

https://github.com/pydoit/doit/commit/b2ed5ba798622d9ac1d2b50de81a4eb24e02f844

This was added by mgorny! :) It also supported tomllib at the same time.
But when someone else packaged it in commit 0e5347e6b3e6d2ac0af7777dfb010368a78dcd06
this was overlooked. Drop the dependency on 3.11+.

Fixes: 0e5347e6b3e6d2ac0af7777dfb010368a78dcd06
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>