app-text/spellutils: add patch to successfully eautoreconf
It uses a heavily vintage autoconf 2.13 in the dist tarball, while
simultaneously packed with gettextize bombs. It fails to configure under
Modern C compilers. Simply regenerating the build system solves all
this.

Bonus: there's a patch that patches a gettext input, which is no longer
used or needed. It dates back to at least the import of ::gentoo into
git.

Closes: https://bugs.gentoo.org/871537
Closes: https://bugs.gentoo.org/906253
Closes: https://github.com/gentoo/gentoo/pull/37307
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>