sys-firmware/intel-microcode: add ucode image generator hooks
This creates a intel microcode image when installing the kernel and when
installkernel/kernel-install is configured to use an initramfs generator other
then dracut. Dracut bundles the microcode in the image it generates, many
other initramfs generators do not and we want these users to also get the
latest CPU microcode.

installkernel-gentoo.git has recently gained support for handling these ucode
images and passing them on to bootloaders and UKI generators.

Note that this should be here and not in installkernel-gentoo.git because we
only want to install this when the initramfs flag is enabled to avoid the
situation where users who don't want an ucode.img get one anyway.

Note that we don't have to worry about MICROCODE_SIGNATURES and
MICROCODE_BLAKCLIST since this selection is taken into account when installing
the microcode to /lib/firmware, at run-time when we read the files from there
we therefore automatically get only the files we want.

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>