sci-libs/symengine: add 0.12.0
This version uses flint-3.x, which includes the package formerly known
as sci-mathematics/arb. As a result we can drop the arb dependency;
and when this version is stable, we can mask sci-mathematics/arb.

Upstream still offers -DWITH_ARB, but plan to remove it soon,

  https://github.com/symengine/symengine/issues/2014

so I have removed it from the ebuild ahead of time. The use of arb is
now goverened by USE=flint.

Signed-off-by: Michael Orlitzky <mjo@gentoo.org>