dev-debug/gdb: support for AMD GPU debugging via USE=rocm
This adds ability to debug AMD GPU code with recently added `dev-libs/rocdbgapi` library.
This feature was introduced in GDB 13 and documented at:
* https://sourceware.org/gdb/current/onlinedocs/gdb.html/AMD-GPU.html
* https://sourceware.org/git/?p=binutils-gdb.git;a=commit;h=18b4d0736bc570c6d2e3e5f6ebc2ad4617d93847

As described in binutils-gdb commit message, this feature requires `--enable-targets=all`,
and if no `--with-amd-dbgapi=yes/no` specified, it tries to probe for library
and enable if found (which is undesirable for Gentoo), therefore it is better to set this flag explicitly.

Signed-off-by: Sv. Lockal <lockalsash@gmail.com>