net-analyzer/notus-scanner: fix overdependency on tomli
Since 22.5.0, it has supported using the stdlib tomllib when available:

https://github.com/greenbone/notus-scanner/commit/73f429133f0f07855d4d0731b6547bc8d672b482

The versioned dependency is poetry semver paranoia. tomli is a stdlib
backport and will remain compatible with the stdlib (presumably by never
bumping to "semver major 3"). But don't argue since it was already in
the ebuild.

In principle, this overdependency could be fixed for 22.6.2 as well, but
going through stabilization for that isn't worth it. This version is
upcoming anyway.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>