dev-util/ruff: add 0.4.10
Includes a backport for a test failure that I originally hit when trying
to package this version, which only manifests because we build with
release mode.

Thanks, debug asserts with side effects. :)

Also the testsuite has introduced the use of the zstd-sys crate. It's
not used at runtime though. Either way, depend on it so that the
testsuite builds faster.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>