profiles/desc: reword and update nginx_modules_http.desc
The wording used in nginx_modules_http.desc to describe the USE flags
was rather inconsistent with how USE flags are described in Gentoo.

This commit features the following changes:

    * The missing descriptions were added for the "ssl"[1],
      "upstream_random", "v2"[1] and "v3"[1] modules.

    * "This module does something" was changed to "Add support for
      something" en masse, as is done elsewhere in Gentoo.

    * The descriptions were brought in line with upstream[2] ones. For
      many of them, the wording was modified to better (sometimes more
      concisely) reflect their function.

    * Update copyright and actually sort the modules!

[1]: Previously, these modules were not USE_EXPAND flags but rather USE
flags on their own. With the introduction of nginx.eclass, they become
NGINX_MODULES_HTTP USE_EXPAND flags, as the rest of NGINX modules are.
[2]: See "Module reference" at the bottom of https://nginx.org/en/docs/

Signed-off-by: Zurab Kvachadze <zurabid2016@gmail.com>