media-video/pipewire: fix file conflicts with pulseaudio-daemon
We don't *have* to be replacing it as the sound server, that's why a USE
flag exists for precisely that. But pipewire 1.2.0+ offers to install
files from pulseaudio-daemon on the theory it can be replacing it...
which means we need to tell it not to do so when USE="-sound-server".

Previously, we installed this whenever gsettings was enabled, even
though its purpose was to enable API usage independent of being the main
sound server.

Closes: https://bugs.gentoo.org/935139
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>