dev-games/clanlib: fix ignored USE=static-libs in slot 2.3
It is a no-op. In gentoo-historical-2.git, the addition of version 2.4.3
to the tree reworked the econf flags and dropped the functionality of
actually passing $(use_enable) based on this USE flag.

https://gitweb.gentoo.org/archive/repo/gentoo-2.git/commit/?id=dde0b5c459de07f06abeea223a9fe5612bdc0ab1

One might ponder why the flag should exist at all, if no one used it. In
fact, due to EAPI 7 it was always installed. :( That being said, the
other slots for this package supported it correctly, which means by
default they weren't installed.

Fix this slot to act the same way: correctly.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>