dev-util/maturin: skip failing test_macosx_deployment_target on sparc
Been failing since rust-1.74 likely due to how it does something
different for >=1.74, and then there is no reason for this to block
stable given 1.4.0 is affected as well. Plus that a macosx deploy test
fails seems unlikely to affect building revdeps with the pep517 backend
(seems unlikely anyone uses it for more than that on sparc).

Bug: https://bugs.gentoo.org/934573
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>