sys-cluster/pmix: enforce the choice of event library
pmix can be built with either libevent or libev, and we need to depend
on the correct one. We depend on libevent. If, somehow, the configure
check is broken, it will "happily" attempt to automagically build against libev
instead, with varying degrees of success.

Make sure it knows which one it MUST pick.

Also update to use modern econf passing style with a local array, so
that comments are supported!

Closes: https://bugs.gentoo.org/884883
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>