sys-libs/fts-standalone: stop renaming .pc file
Stop renaming musl-fts.pc -> fts-standalone.pc. While it might be more
pure as a name (as it might be useful on other libcs), it's not within our
gift to rename this and it has ecosystem-wide effects.

Other distributions are using the upstream name.

Also, install a symlink for the old fts-standalone name to keep broken
checks working.

While here, fix MisplacedWeakBlocker too.

Closes: https://bugs.gentoo.org/895946
Signed-off-by: Sam James <sam@gentoo.org>