media-video/pipewire: revbump recently modified ebuilds
I submitted a series of fixes for automagic dependencies etc. Most of
them were merged, but one of them ("fix automagic dependency on
libpulse") was deferred since we are not yet sure the best way to do it.

Unfortunately, that was also the first patch in the series, and it had
the revbump for the changes we made. When skipping that patch, git ended
up modifying the existing versions in place which means existing
automagic dependencies didn't trigger a rebuild. Add that revbump in
now.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>