dev-qt/qtmultimedia: disable alsa when pulseaudio is enabled (qt6)
Could revisit enabling both if qtmultimedia tries higher-end
backends like pulseaudio first in the future, but currently
it is both marked experimental and without autodetect:

    qt_feature("alsa" PUBLIC PRIVATE
        LABEL "ALSA (experimental)"
        AUTODETECT false

So only enable it if it's the only option we have.

Have not checked the state of this for Qt5, but probably better
off not touching at this point unless someone has problems
(it's possible the audio sink was going through gstreamer
preventing issues).

Closes: https://bugs.gentoo.org/935146
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>