dev-qt/qt-creator: update live
Not tested much, but did build successfully once w/ Go bits.

May or may not consider making this optional when there is a release
after a closer look. Does not have a proper option, but it already
skips it if it does not find Go... so -DGO_BIN=no or so would likely
work.

Being off by default would also be less annoying for go rebuilds,
I doubt anyone wants to rebuild something massive like qt-creator
every go bump. Not that I really checked what this is used for yet.

For now not going to touch this too much given it may get changed
around upstream.

Closes: https://bugs.gentoo.org/935332
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>