dev-qt/qtwebengine: backport wayland text rendering fix (qt6)
See patch header.

Tempting to revbump, albeit will still spare users for now. Will
at least give stable users the fixed vresion when stabilized soon.
~testing users annoyed by this are invited to manually rebuild.

Also add to 6.7.9999 so it is not forgotten with 6.7.3, "currently"
not marked to be picked to the 6.7 branch for the 6.7.3 release
and may or may not still be needed until 6.8.0.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>