app-shells/bash: add 5.2_p26-r7 with XTWINOPS support
Xterm is able to push and pop window titles to a stack and there are a
few other terminal emulators that can do so (alacritty and foot being
the only ones that I am aware of at the present time). Take advantage of
this feature so as to reinstate automatic window title setting in the
case that the PTY is owned by sshd(8). Unfortunately, there are a lot of
terminal emulators that falsely advertise themselves as being
xterm-compatible, making it impossible to reliably identify xterm
itself. However, we can reliably identify alacritty and foot so let's
support those two to begin with.

Have the genfun_set_win_title function export a variable named
SHELL_SETS_TITLE upon the first occasion that it is called. Presently,
nothing responds to this variable but the intention is to eventually
have portage respond to it. Portage implements heuristics and behaviours
that are horrifyingly broken. For instance, it considers the mere
presence of PROMPT_COMMAND as somehow proving that the interactive shell
uses it for nothing other than to set the title, despite the fact that:

- the contents of PROMPT_COMMAND may be arbitrarily defined by the user
- the purpose of PROMPT_COMMAND is whatever the user may wish it to be
- nobody in their right mind would export PROMPT_COMMAND
- PROMPT_COMMAND can be an array since 5.1 (making it unexportable)

Worse still, in the event that portage is somehow able to ascertain the
value of PROMPT_COMMAND, it takes its first element and proceeds to
inject its value into an invocation of either sh, $SHELL or bash -c,
irrespective of the consequences. No, I'm not making this up.

As such, the purpose of the SHELL_SETS_TITLE variable is to act as a
straightforward indicator that an interactive shell exists as an
ancestor process and that it will take it upon itself to set a fresh
window title upon its primary prompt being displayed.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>