dev-qt/qt-creator: unconditionally depend on yaml-cpp
Missed that a plugin (qtapplicationmanager) added a dependency on
yaml-cpp while comparing comparing qt-creator-12 -> 13 and is no
longer clang-only.

Feel it's not worth a USE for the plugin on such a big package
(not for minimalism), so depending unconditionally unless someone
really wants to avoid it (at best it could be interesting to avoid
:= rebuilds).

Technically needs a revbump, but qt-creator takes a while to build
and most users already were pulling it through IUSE=+clang.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>