net-analyzer/greenbone-feed-sync: revert "drop config file patching in src_test"
This essentially reverts commit 0375a7b2b589d7ffe2b73a3426d8c0abd4f4b38e, but
adjusts the sed comment to remove the reference to sandbox violations,
because this are not what we are seeing without the sed.

We also now use T to store config.py.orig, as it brings down the line
length.

Furthermore, we keep the "${WORKDIR}/${P}"/ part dropped, as it is
unnecessary.

Closes: https://bugs.gentoo.org/932836
Signed-off-by: Florian Schmaus <flow@gentoo.org>