dev-python/setuptools: Add a PDEP on dev-python/trove-classifiers
Modern versions of setuptools optionally use
dev-python/trove-classifiers for verification.  This has the side issue
that if an old version happens to be installed, random packages may
fail to install.  Until upstream makes this properly optional (and we
disable it), depend on the very newest version.

Andrew noticed it while trying to bump dev-python/pyzotero.

Signed-off-by: Michał Górny <mgorny@gentoo.org>