www-client/chromium: add 127.0.6533.26
This release brings some major changes to the way that Chromium
is packaged in Gentoo. After masking GCC support since it broke beyond
the capabilities of the maintainers to fix in 120, changes
in 127 have made dependencies on LLVM stronger: in addition to
requiring LLVM for Rust we now directly depend on LLVM libraries for
bindgen.

As a result we're experimenting with completely dropping GCC support
from the ebuild. This isn't a decision that was made lightly, however
the advantages to going down this path make it very appealing:

- Our builds are closer to upstream (and nominally supportable)
- We don't need to maintain so many out-of-tree patches
  + We may actually be able to drop them entirely for most arches
- Ebuild logic is significantly simplified, making it easier to review
- Decreased maintainer hair-pulling.

The removed GCC support could be revisited if there is sufficient
interest, and as an added benefit the new, streamlined, and better
documented ebuild should provide a better basis for maintaining it.

One of the outcomes of this decision to double-down on LLVM support has
been the opportunity to properly select an implementation via upgrading
to the `llvm-r1` eclass, which also enables us to start trying to make
the Rust->LLVM dependency a bit more clear for our ebuild.

Notable changes:

- `llvm-r1` eclass
- Dropped GCC support
- Always bundled libcxx for stdlib
- GN/bindgen wrapper updated to work with system bindgen
  + needs upstreaming

Signed-off-by: Matt Jolly <kangie@gentoo.org>