Revert "tree-sitter-grammar.eclass: extended packaging"
This reverts commit fbe4bc7b9dc970271127a5f167c95da78e32e282.

The previous change does not install the NeoVim symlinks correctly as they
point to /usr/lib/tree-sitter/liblang.so instead of /usr/lib/liblang.so.

Also the query files are not installed in per-package directories and have
common names (e.g. highlight.scm) which is causing file collisions.

To be fixed in a follow up commit which will introduce
tree-sitter-grammar-r1.eclass, so that ebuilds can be rev-bumped so that
users can rebuild and get the changes.

Bug: https://bugs.gentoo.org/935631
Bug: https://bugs.gentoo.org/935641
Signed-off-by: Matthew Smith <matthew@gentoo.org>