dev-build/cmake: mark as LTO-unsafe under bootstrap as well
We filter-lto due to ODR violations. But only in src_configure, which is
common enough as that's where we set up the environment for configuring,
in general. However cmake is a bit special as we also need to bootstrap
it, and we do this in src_prepare(). We need to do that filtering there
as well.

While we are at it, the solaris linking issue is most probably relevant
the same way, although I have not actually tested this...

Bug: https://bugs.gentoo.org/858335
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>
Acked-by: Sam James <sam@gentoo.org>