net-misc/yt-dlp: add 2024.07.07
Another minor update, this time for a potential vulnerability
(currently no CVE assigned), may as well straight-to-stable too.

Technically not affected given we do not package phantomjs anymore
and it been gone long before yt-dlp was in the tree. Even if a
user had this anyway, they would have to use the extractor from a
lesser used site and it is not believed that yt-dlp+phantomjs has
been targeted (more of a just-in-case fix).

...so skipping a Gentoo security bug this time around.

Odds are that there are more potential issue if have PhantomJS
installed, ancient no-longer-updated javascript runner is going
to inevitably be a security hazard.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>