net-dns/opendnssec: add patch to build with gcc 14
Upstream's website appears to be having... issues. I did find a bug
reporting link but it led to a subdomain with broken SSL. Unclear how to
progress further.

The sum totality of the issues is a lot of missing headers. Not only did
this affect both configure checks as well as the installable artifacts,
but it also triggered LTO type mismatch errors when the missing stdlib
headers resulted in fake guessed prototypes that then mismatched libc.
Killing two birds with one stone!

Closes: https://bugs.gentoo.org/871294
Closes: https://bugs.gentoo.org/879727
Closes: https://bugs.gentoo.org/924367
Closes: https://bugs.gentoo.org/927176
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>