sci-physics/pythia: mark as LTO-unsafe
It's very old code. So old, in fact, that upstream doesn't support it
anymore and introduced a major new release series (which resulted in
adding an outright SLOT to the ebuild) to port it to C++. And that was
in 2012.

The error itself is, naturally, fortran related. :) :) So there is
simply no point in reporting this anywhere. The newer SLOT works fine.
Simply prevent LTO. Anyways, no one is using 6.x except "to allow older
published results to be reproduced, should the need arise". Free
permission to not care about LTO...

Closes: https://bugs.gentoo.org/927728
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>