x11-libs/libgxim: backport gcc 14 fix for implicit-function-decl
Upstream has fixed this in 2018 but simply not released a new version:
https://bitbucket.org/tagoh/libgxim/commits/274976ff39d9142169a18655f5013e66233ac9c8

Due to the implicit function declaration, compilation fails on gcc 14.
On gcc 13, instead, there is an lto-type-mismatch because the K&R C
style incorrect prototype was... incorrect and mismatched the required
types. :)

The same upstream commit also includes an unrelated fix for a different
build issue: -Werror=format-security.

Closes: https://bugs.gentoo.org/875056
Closes: https://bugs.gentoo.org/928090
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>