profiles/updates: update historic move of a package to its current location
We moved {sci-libs -> dev-python}/rtree, but once upon a time we had
also moved it *to* that starting location, for capitalization reasons.
Now pkgcheck throws an error that it is a "MultiMovePackageUpdate":

> 'sci-libs/Rtree': multi-move update: sci-libs/Rtree -> sci-libs/rtree -> dev-python/rtree

Assuming anyone actually does have it installed from 2020, we should
still move it to the current location, not a transition location. The
recent move didn't actually take this into account, and broke the
repo-qa CI.

Fixes: 287f462eba67dec7036cd51c4372896144219184
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>