dev-build/meson: skip another test that inexplicably fails
I do not know why clippy-driver fails to compile sanity.rs, but it does.
It works flawlessly outside of portage, and is tested like that in
upstream CI running on Gentoo. And tests pass on my machine, in a git
checkout of meson. But not when running the same thing in portage. Even
manually constructing a sanity.rs file in ${FILESDIR} and executing
`clippy-driver ...` in src_test fails, but the exact same command
run inside of `sudo -u portage bash` upon the same file, succeeds.

It's not the portage sandbox, but it is something about the portage
environment, and I have not been able to figure out what. And it's
making it impossible to run the testsuite.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>