media-libs/mesa: update NVK rust crate handling to use meson packagecache
We do a bit of a song and a dance to slip rust crates into meson's
subprojects directory and migrate the meson.build files over. But meson
can handle this builtin via:

- `meson subprojects download`
- providing the tarball seeds in the packagecache and relying on
  extraction at time of use

It's a bit less fragile than manual extraction because meson guarantees
that patch_directory handling is done and no need to manually
reimplement it in bash. The downside, for the moment, is that previously
the checksums didn't actually match what the meson wraps expected, due
to the differences between crates.io distfiles and github distfiles. So
we cannot rename the files on the mirror to what we want.

Once all the crate versions have been bumped by the passing of time, we
can rename them in SRC_URI and simply use ${MESON_PACKAGE_CACHE_DIR},
new since meson 1.3.0 (we already require 1.4 so that's okay), to tell
meson that it should use wrap tarballs from ${DISTDIR} without any
intervention.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>